requirements/common.txt
# Django
Django>=1.8,<1.8.99

# DJ Database
dj-database-url==0.3.0

# Postgresql
psycopg2==2.5.2

# Django Extensions
django-extensions==1.5.2
six==1.6.1
shortuuid==0.4.2

# IPython
ipython==2.0.0

# Pillow (PIL)
Pillow==2.7.0

# Django Storages
# django-storages==1.1.8

# Boto (for Amazon S3)
# boto==2.27.0

# Django Rest Framework
djangorestframework==3.0.4

# CORS access as the server
django-cors-headers==0.13

wsgiref==0.1.2

# Celery
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.19
celery==3.1.17
django-celery==3.1.16
kombu==3.0.24
pytz==2014.10

flower==0.7.3
tornado==4.0.2
certifi==14.05.14
backports.ssl_match_hostname==3.4.0.2

# Gunicorn
gunicorn==18.0

